django/requirements.txt
apipkg==1.5
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
attrs==20.3.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coverage==5.3
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-environ==0.4.5
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
execnet==1.7.1
Faker==0.9.1
html5lib==1.0.1
idna==2.8
iniconfig==1.1.1
ipaddr==2.2.0
isort==5.6.4
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
mixer==6.1.3
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
pluggy==0.13.1
progress==1.5
psycopg2-binary==2.8.6
py==1.9.0
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.6
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-forked==1.3.0
pytest-sugar==0.9.4
pytest-xdist==2.1.0
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
termcolor==1.1.0
text-unidecode==1.2
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
